Separate components and refactor
