+++ src/components/DataTable.tsx
@@ -1,275 +1,26 @@
-import {  
+import React, { 
+  MouseEvent, 
   useEffect, 
-  useState, 
-  MouseEvent,
-  ChangeEvent
+  useState 
 } from 'react';
 import axios from "axios";
-import { useTheme, styled } from '@mui/material/styles';
-import Box from '@mui/material/Box';
 import TableSortLabel from "@mui/material/TableSortLabel";
 import Table from '@mui/material/Table';
-import Collapse from "@mui/material/Collapse";
 import TableBody from '@mui/material/TableBody';
 import TableHead from "@mui/material/TableHead";
-import TableCell, { tableCellClasses } from '@mui/material/TableCell';
+import TableCell from '@mui/material/TableCell';
 import TableContainer from '@mui/material/TableContainer';
 import TableFooter from '@mui/material/TableFooter';
 import TablePagination from '@mui/material/TablePagination';
-import TableRow from '@mui/material/TableRow';
 import Paper from '@mui/material/Paper';
-import IconButton from '@mui/material/IconButton';
-import FirstPageIcon from '@mui/icons-material/FirstPage';
-import KeyboardArrowLeft from '@mui/icons-material/KeyboardArrowLeft';
-import KeyboardArrowRight from '@mui/icons-material/KeyboardArrowRight';
-import LastPageIcon from '@mui/icons-material/LastPage';
-import Typography from "@mui/material/Typography";
-import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
-import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
-
-interface TablePaginationActionsProps {
-  count: number;
-  page: number;
-  rowsPerPage: number;
-  onPageChange: (
-    event: MouseEvent<HTMLButtonElement>,
-    newPage: number,
-  ) => void;
-}
-
-type addressType = {
-  street: string,
-  suite: string,
-  city: string,
-  zipcode: number,
-  geo: {
-    lat: string,
-    lng: string
-  }
-}
-
-type companyType = {
-  name: string,
-  catchPhrase: string,
-  bs: string
-}
-
-type userType = {
-  id: number,
-  name: string, 
-  username: string, 
-  email: string,
-  address: addressType
-  phone: string, 
-  website: string,
-  company: companyType
-}
-
-function TablePaginationActions(props: TablePaginationActionsProps) {
-  const theme = useTheme();
-  const { count, page, rowsPerPage, onPageChange } = props;
-
-  const handleFirstPageButtonClick = (
-    event: MouseEvent<HTMLButtonElement>,
-  ) => {
-    onPageChange(event, 0);
-  };
-
-  const handleBackButtonClick = (event: MouseEvent<HTMLButtonElement>) => {
-    onPageChange(event, page - 1);
-  };
-
-  const handleNextButtonClick = (event: MouseEvent<HTMLButtonElement>) => {
-    onPageChange(event, page + 1);
-  };
-
-  const handleLastPageButtonClick = (event: MouseEvent<HTMLButtonElement>) => {
-    onPageChange(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-  };
-
-  return (
-    <Box sx={{ flexShrink: 0, ml: 2.5 }}>
-      <IconButton
-        onClick={handleFirstPageButtonClick}
-        disabled={page === 0}
-        aria-label="first page"
-      >
-        {theme.direction === 'rtl' ? <LastPageIcon /> : <FirstPageIcon />}
-      </IconButton>
-      <IconButton
-        onClick={handleBackButtonClick}
-        disabled={page === 0}
-        aria-label="previous page"
-      >
-        {theme.direction === 'rtl' ? <KeyboardArrowRight /> : <KeyboardArrowLeft />}
-      </IconButton>
-      <IconButton
-        onClick={handleNextButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="next page"
-      >
-        {theme.direction === 'rtl' ? <KeyboardArrowLeft /> : <KeyboardArrowRight />}
-      </IconButton>
-      <IconButton
-        onClick={handleLastPageButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="last page"
-      >
-        {theme.direction === 'rtl' ? <FirstPageIcon /> : <LastPageIcon />}
-      </IconButton>
-    </Box>
-  );
-}
-
-function createData(
-  id: number,
-  name: string,
-  username: string,
-  email: string,
-  address: addressType,
-  phone: string, 
-  website: string,
-  company: companyType
-  ) {
-    return {
-      id,
-      name,
-      username,
-      email,
-      address,
-      phone, 
-      website,
-      company
-  };
-}
-
-function Row(props: { row: ReturnType<typeof createData> }) {
-  const { row } = props;
-  const [openAddress, setOpenAddress] = useState(false);
-  const [openCompany, setOpenCompany] = useState(false);
-
-  return (
-    <>
-      <StyledTableRow sx={{ "& > *": { borderBottom: "unset" } }}>
-
-        <StyledTableCell component="th" scope="row">
-          {row.id}
-        </StyledTableCell>
-        <StyledTableCell align="right" style={{ width: 225 }}>{row.name}</StyledTableCell>
-        <StyledTableCell align="right" style={{ width: 225 }}>{row.username}</StyledTableCell>
-        <StyledTableCell align="right" style={{ width: 225 }}>{row.email}</StyledTableCell>
-        <StyledTableCell align="right" style={{ width: 75 }}>
-          <IconButton
-            aria-label="expand row"
-            size="small"
-            onClick={() => setOpenAddress(!openAddress)}
-          >
-            {openAddress ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-          </IconButton>
-        </StyledTableCell>
-        <StyledTableCell align="right" style={{ width: 225 }}>{row.phone}</StyledTableCell>
-        <StyledTableCell align="right" style={{ width: 225 }}>{row.website}</StyledTableCell>
-        <StyledTableCell align="right" style={{ width: 75 }}>
-          <IconButton
-            aria-label="expand row"
-            size="small"
-            onClick={() => setOpenCompany(!openCompany)}
-          >
-            {openCompany ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-          </IconButton>
-        </StyledTableCell>
-      </StyledTableRow>
-      <StyledTableRow>
-        <StyledTableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
-          <Collapse in={openAddress} timeout="auto" unmountOnExit>
-            <Box sx={{ margin: 1 }}>
-              <Typography variant="h6" gutterBottom component="div">
-                Address
-              </Typography>
-              <Table size="small" aria-label="purchases">
-                <TableHead>
-                  <StyledTableRow>
-                    <StyledTableCell>Street</StyledTableCell>
-                    <StyledTableCell align="right">Suite</StyledTableCell>
-                    <StyledTableCell align="right">City</StyledTableCell>
-                    <StyledTableCell align="right">Zipcode</StyledTableCell>
-                    <StyledTableCell align="right">Geolocation</StyledTableCell>
-                  </StyledTableRow>
-                </TableHead>
-                <TableBody>
-                  <StyledTableRow>
-                    <StyledTableCell component="th" scope="row">
-                      {row.address.street}
-                    </StyledTableCell>
-                    <StyledTableCell align="right" style={{ width: 200 }}>{row.address.suite}</StyledTableCell>
-                    <StyledTableCell align="right" style={{ width: 200 }}>{row.address.city}</StyledTableCell>
-                    <StyledTableCell align="right" style={{ width: 200 }}>{row.address.zipcode}</StyledTableCell>
-                    <StyledTableCell align="right" style={{ width: 250 }}><strong>Lat: </strong>{row.address.geo.lat}
-                      <b>, Long: </b>{row.address.geo.lng}</StyledTableCell>
-                  </StyledTableRow>
-                </TableBody>
-              </Table>
-            </Box>
-          </Collapse>
-        </StyledTableCell>
-      </StyledTableRow>
-
-      <StyledTableRow>
-        <StyledTableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
-          <Collapse in={openCompany} timeout="auto" unmountOnExit>
-            <Box sx={{ margin: 1 }}>
-              <Typography variant="h6" gutterBottom component="div">
-                Company
-              </Typography>
-              <Table size="small" aria-label="purchases">
-                <TableHead>
-                  <StyledTableRow>
-                    <StyledTableCell>Name</StyledTableCell>
-                    <StyledTableCell align="right">Catch Phrase</StyledTableCell>
-                    <StyledTableCell align="right">BS</StyledTableCell>
-                  </StyledTableRow>
-                </TableHead>
-                <TableBody>
-                  <StyledTableRow>
-                    <StyledTableCell component="th" scope="row">
-                      {row.company.name}
-                    </StyledTableCell>
-                    <StyledTableCell align="right">{row.company.catchPhrase}</StyledTableCell>
-                    <StyledTableCell align="right">{row.company.bs}</StyledTableCell>
-                  </StyledTableRow>
-                </TableBody>
-              </Table>
-            </Box>
-          </Collapse>
-        </StyledTableCell>
-      </StyledTableRow>
-    </>
-  );
-}
-
-const StyledTableCell = styled(TableCell)(({ theme }) => ({
-  [`&.${tableCellClasses.head}`]: {
-    backgroundColor: theme.palette.common.black,
-    color: theme.palette.common.white,
-  },
-  [`&.${tableCellClasses.body}`]: {
-    fontSize: 14,
-  },
-}));
-
-const StyledTableRow = styled(TableRow)(({ theme }) => ({
-  '&:nth-of-type(odd)': {
-    backgroundColor: theme.palette.action.hover,
-  },
-  // hide last border
-  '&:last-child td, &:last-child th': {
-    border: 0,
-  },
-}));
+import { userType } from './Table.types'
+import { StyledTableCell, StyledTableRow } from './styledTable';
+import Row from './Row';
+import TablePaginationActions from './TablePaginationActions'
 
 export default function DataTable() {
-  const [page, setPage] = useState(0);
-  const [rowsPerPage, setRowsPerPage] = useState(3);
+  const [page, setPage] = React.useState(0);
+  const [rowsPerPage, setRowsPerPage] = React.useState(3);
   const [rows, setRows] = useState<userType[]>([]);
 
   useEffect(() => {
@@ -296,7 +47,7 @@ export default function DataTable() {
   };
 
   const handleChangeRowsPerPage = (
-    event: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
+    event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>,
   ) => {
     setRowsPerPage(parseInt(event.target.value, 10));
     setPage(0);
@@ -304,7 +55,7 @@ export default function DataTable() {
 
   type orderType = "asc" | "desc";
 
-  const [orderDirection, setOrderDirection] = useState<orderType>("asc");
+  const [orderDirection, setOrderDirection] = React.useState<orderType>("asc");
 
   const sortArray = (arr: userType[], orderBy: orderType) => {
     switch (orderBy) {
@@ -320,21 +71,21 @@ export default function DataTable() {
     }
   };
 
-  const sortAlphaArray = (arr: userType[], orderBy: orderType) => {
-    const sortedData = [...arr].sort((a, b) => {
-      if (orderBy === 'asc') {
-          return a.username.localeCompare(b.username);
-      } else {
-          return b.username.localeCompare(a.username);
-      }
-    });
-    return sortedData;
-  }
-
   const handleSortRequest = () => {
     setRows(sortArray(rows, orderDirection));
     setOrderDirection(orderDirection === "asc" ? "desc" : "asc");
   };
+
+  const sortAlphaArray = (arr: userType[], orderBy: orderType) => {
+    const sortedData = [...arr].sort((a: userType, b: userType) => {
+      if (orderBy === 'asc') {
+        return a.username.localeCompare(b.username);
+      } else {
+         return b.username.localeCompare(a.username);
+      }
+    });
+    return sortedData;
+  }
 
   const handleAlphaSort = () => {
     setRows(sortAlphaArray(rows, orderDirection));
@@ -345,11 +96,11 @@ export default function DataTable() {
     <TableContainer component={Paper}>
       <Table sx={{ minWidth: 500 }} aria-label="custom pagination table">
         <TableHead>
-          <TableRow>
-          <StyledTableCell align="left" onClick={handleSortRequest}>
-            <TableSortLabel active={true} direction={orderDirection}>
-              Id
-            </TableSortLabel>
+          <StyledTableRow>
+            <StyledTableCell align="left" onClick={handleSortRequest}>
+              <TableSortLabel active={true} direction={orderDirection}>
+                Id
+              </TableSortLabel>
             </StyledTableCell>
             <StyledTableCell align="right">Name</StyledTableCell>
             <StyledTableCell align="right" onClick={handleAlphaSort}>
@@ -362,7 +113,7 @@ export default function DataTable() {
             <StyledTableCell align="right">Phone</StyledTableCell>
             <StyledTableCell align="right">Website</StyledTableCell>
             <StyledTableCell align="right">Company</StyledTableCell>
-          </TableRow>
+          </StyledTableRow>
         </TableHead>
         <TableBody>
           {(rowsPerPage > 0
@@ -372,13 +123,13 @@ export default function DataTable() {
             <Row key={row.id} row={row} />
           ))}
           {emptyRows > 0 && (
-            <TableRow style={{ height: 53 * emptyRows }}>
-              <StyledTableCell colSpan={6} />
-            </TableRow>
+            <StyledTableRow style={{ height: 53 * emptyRows }}>
+              <TableCell colSpan={6} />
+            </StyledTableRow>
           )}
         </TableBody>
         <TableFooter>
-          <TableRow>
+          <StyledTableRow>
             <TablePagination
               rowsPerPageOptions={[3, 6, 9, { label: 'All', value: -1 }]}
               colSpan={8}
@@ -395,7 +146,7 @@ export default function DataTable() {
               onRowsPerPageChange={handleChangeRowsPerPage}
               ActionsComponent={TablePaginationActions}
             />
-          </TableRow>
+          </StyledTableRow>
         </TableFooter>
       </Table>
     </TableContainer>
